Module Tezos_client_007_PsDELPH1.Michelson_v1_entrypoints

val script_entrypoint_type : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Script.expr -> entrypoint:string -> Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Script.expr option Tezos_base.TzPervasives.tzresult Lwt.t

Returns Some type if the contract has an entrypoint of type type. None if it does not exists.

val contract_entrypoint_type : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> contract:Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Contract.t -> entrypoint:string -> Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Script.expr option Tezos_base.TzPervasives.tzresult Lwt.t

Returns Some type if the script has an entrypoint of type type. None if it does not exists.

val print_entrypoint_type : Tezos_client_base.Client_context.printer -> ?on_errors: (Tezos_base.TzPervasives.error list -> unit Tezos_base.TzPervasives.tzresult Lwt.t) -> emacs:bool -> ?contract:Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Contract.t -> ?script_name:string -> entrypoint:string -> Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Script.expr option Tezos_base.TzPervasives.tzresult -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val list_contract_unreachables : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> contract:Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Contract.t -> Tezos_protocol_007_PsDELPH1.Protocol.Michelson_v1_primitives.prim list list Tezos_base.TzPervasives.tzresult Lwt.t

List paths of unreachable parameters. Only useful to test the stitching, as no such parameter should be allowed in originated contracts.

val list_unreachables : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Script.expr -> Tezos_protocol_007_PsDELPH1.Protocol.Michelson_v1_primitives.prim list list Tezos_base.TzPervasives.tzresult Lwt.t

val print_unreachables : Tezos_client_base.Client_context.printer -> ?on_errors: (Tezos_base.TzPervasives.error list -> unit Tezos_base.TzPervasives.tzresult Lwt.t) -> emacs:bool -> ?contract:Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Contract.t -> ?script_name:string -> Tezos_protocol_007_PsDELPH1.Protocol.Michelson_v1_primitives.prim list list Tezos_base.TzPervasives.tzresult -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val list_contract_entrypoints : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> contract:Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Contract.t -> (string * Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Script.expr) list Tezos_base.TzPervasives.tzresult Lwt.t

List the contract entrypoints with their types. If their is no explicit default, th type of default entrypoint will still be given.

val list_entrypoints : Protocol_client_context.rpc_context -> chain:Tezos_shell_services.Chain_services.chain -> block:Tezos_shell_services.Block_services.block -> Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Script.expr -> (string * Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Script.expr) list Tezos_base.TzPervasives.tzresult Lwt.t

List the script entrypoints with their types.

val print_entrypoints_list : Tezos_client_base.Client_context.printer -> ?on_errors: (Tezos_base.TzPervasives.error list -> unit Tezos_base.TzPervasives.tzresult Lwt.t) -> emacs:bool -> ?contract:Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Contract.t -> ?script_name:string -> (string * Tezos_protocol_007_PsDELPH1.Protocol.Alpha_context.Script.expr) list Tezos_base.TzPervasives.tzresult -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Print the contract entrypoints with their types.